+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased 5.0.1]
+
+- Issue #249: Constructors in test classes were instrumented as test methods, producing uncompilable code.
+
 ## [5.0.0] 2026-07-16
 
 ### Added